# Bump python, small image size, close python CVE, add new flag UVICORN_SSL_CA_TYPE