--- a/.env.example
+++ b/.env.example
@@ -10,6 +10,7 @@ UVICORN_PORT = 8000
 # UVICORN_UDS: "/run/marzban.socket"
 # UVICORN_SSL_CERTFILE = "/var/lib/marzban/certs/example.com/fullchain.pem"
 # UVICORN_SSL_KEYFILE = "/var/lib/marzban/certs/example.com/key.pem"
+# UVICORN_SSL_CA_TYPE = "public"
 
 # DASHBOARD_PATH = "/dashboard/"