--- a/Dockerfile
+++ b/Dockerfile
@@ -1,21 +1,28 @@
-FROM python:3.10-slim
+FROM python:3.10-slim AS build
 
 ENV PYTHONUNBUFFERED 1
 
 WORKDIR /code
 
 RUN apt-get update \
-    && apt-get install -y curl unzip gcc python3-dev \
+    && apt-get install -y --no-install-recommends curl unzip gcc python3-dev \
+    && curl -L https://github.com/Gozargah/Marzban-scripts/raw/master/install_latest_xray.sh | bash \
     && rm -rf /var/lib/apt/lists/*
 
-RUN bash -c "$(curl -L https://github.com/Gozargah/Marzban-scripts/raw/master/install_latest_xray.sh)"
-
 COPY ./requirements.txt /code/
-RUN pip install --no-cache-dir --upgrade -r /code/requirements.txt
+RUN python3 -m pip install --upgrade pip setuptools \
+    && pip install --no-cache-dir --upgrade -r /code/requirements.txt
+
+FROM python:3.10-slim
+
+WORKDIR /code
+
+RUN rm -rf /usr/local/lib/python3.10/site-packages/*
+
+COPY --from=build /usr/local/lib/python3.10/site-packages /usr/local/lib/python3.10/site-packages
+COPY --from=build /usr/local/bin /usr/local/bin
 
 COPY . /code
-
-RUN apt-get remove -y curl unzip gcc python3-dev
 
 RUN ln -s /code/marzban-cli.py /usr/bin/marzban-cli \
     && chmod +x /usr/bin/marzban-cli \