--- a/requirements.txt
+++ b/requirements.txt
@@ -12,7 +12,7 @@ commentjson==0.9.0
 cryptography==43.0.1
 Deprecated==1.2.13
 ecdsa==0.19.0
-fastapi==0.96.0
+fastapi==0.115.2
 greenlet==2.0.1
 grpcio==1.56.2
 grpcio-tools==1.44.0
@@ -62,9 +62,9 @@ rsa==4.9
 six==1.16.0
 sniffio==1.3.0
 SQLAlchemy==1.4.44
-starlette==0.27.0
+starlette==0.40.0
 typer==0.7.0
-typing_extensions==4.4.0
+typing_extensions==4.8.0
 tzdata==2022.6
 tzlocal==4.2
 urllib3==1.26.19