**What's done:**

* Reduced the size of the Docker image by almost half (47%). Build optimization.

```
❯ docker images | grep gozargah/marzban
gozargah/marzban    dev-0911    b9caa77a3b93      About a minute ago        361MB
gozargah/marzban    latest      c0bf4f4f415b      3 weeks ago               682MB
```

* Closing vulnerable CVE python packages.

_before:_ 
```
gozargah/marzban:latest (debian 12.7)
Total: 983 (UNKNOWN: 37, LOW: 257, MEDIUM: 542, HIGH: 143, CRITICAL: 4)

Python (python-pkg)
Total: 42 (UNKNOWN: 0, LOW: 4, MEDIUM: 20, HIGH: 17, CRITICAL: 1)
```

_After updating the libraries and docker image:_

```
gozargah/marzban:dev-0911 (debian 12.7)
Total: 102 (UNKNOWN: 0, LOW: 74, MEDIUM: 17, HIGH: 10, CRITICAL: 1)

Python (python-pkg)

Total: 2 (UNKNOWN: 0, LOW: 0, MEDIUM: 1, HIGH: 1, CRITICAL: 0)

┌────────────────────────┬────────────────┬──────────┬──────────┬───────────────────┬───────────────┬──────────────────────────────────────────────────────────────┐
│        Library         │ Vulnerability  │ Severity │  Status  │ Installed Version │ Fixed Version │                            Title                             │
├────────────────────────┼────────────────┼──────────┼──────────┼───────────────────┼───────────────┼──────────────────────────────────────────────────────────────┤
│ python-jose (METADATA) │ CVE-2024-33663 │ HIGH     │ affected │ 3.3.0             │               │ python-jose: algorithm confusion with OpenSSH ECDSA keys and │
│                        │                │          │          │                   │               │ other key formats                                            │
│                        │                │          │          │                   │               │ https://avd.aquasec.com/nvd/cve-2024-33663                   │
│                        ├────────────────┼──────────┤          │                   ├───────────────┼──────────────────────────────────────────────────────────────┤
│                        │ CVE-2024-33664 │ MEDIUM   │          │                   │               │ python-jose: allows attackers to cause a denial of service   │
│                        │                │          │          │                   │               │ https://avd.aquasec.com/nvd/cve-2024-33664                   │
└────────────────────────┴────────────────┴──────────┴──────────┴───────────────────┴───────────────┴──────────────────────────────────────────────────────────────┘
```

* Added a new flag `UVICORN_SSL_CA_TYPE` to enable the use of a self-signed certificate

```
ubuntu@main:/opt/marzban$ cat .env | grep UVICORN_SSL_CA_TYPE
UVICORN_SSL_CA_TYPE = "private

INFO  [alembic.runtime.migration] Running upgrade a9cfd5611a82 -> 2ea33513efc0, noise for sqlite
INFO  [alembic.runtime.migration] Running upgrade 2ea33513efc0 -> 21226bc711ac, add threshold to NotificationReminder

IMPORTANT!
You're running Marzban with: UVICORN_SSL_CA_TYPE: private.
Self-signed CAs are useful in testing or internal use cases, they’re not suitable for secure public internet communications.

INFO:     Started server process [1]
INFO:     Waiting for application startup.
INFO:     Generating Xray core config
INFO:     Xray core config generated in 0.03 seconds
INFO:     Starting main Xray core
WARNING:  Xray core 24.11.5 started
INFO:     Starting nodes Xray core
INFO:     Application startup complete.
INFO:     Uvicorn running on https://0.0.0.0:10000 (Press CTRL+C to quit)
INFO:     127.0.0.1:38658 - "POST /api/admin/token HTTP/1.1" 200 OK
```

delete UVICORN_SSL_CA_TYPE from .env

Getting valid error about self CA
```
ValueError: Certificate verification failed: The certificate is self-signed and not issued by a trusted CA.
```

* Added venv to `.dockerignore`

* Update python version to 3.12 and some dependency libs

**Check functionality:**

                
<img width="604" alt="image" src="https://github.com/user-attachments/assets/d1c1b0f3-428c-4c38-9e8d-8909d12fccf8">
<img width="1771" alt="image" src="https://github.com/user-attachments/assets/29ada8fc-e991-4a47-94c4-d4bb130deba7">
<img width="437" alt="Using Marzban-Node, you are able to scale up your" src="https://github.com/user-attachments/assets/c6ae1c7c-88b4-4dc2-b8df-2070e43326fa">


```
ubuntu@main:~$ docker ps
CONTAINER ID   IMAGE                       COMMAND                  CREATED              STATUS                        PORTS     NAMES
bb414393a96b   haproxy:2.4.25              "docker-entrypoint.s…"   About a minute ago   Up 49 seconds                           marzban-haproxy-1
1b5a7e5c6510   vladkmrdnv/marzban:latest   "bash -c 'alembic up…"   About a minute ago   Up 49 seconds                           marzban-marzban-1
e4a1016a89be   mariadb:lts                 "docker-entrypoint.s…"   About a minute ago   Up About a minute (healthy)             marzban-mariadb-1
```

```
INFO:     Started server process [1]
INFO:     Waiting for application startup.
INFO:     Generating Xray core config
INFO:     Xray core config generated in 0.04 seconds
INFO:     Starting main Xray core
WARNING:  Xray core 24.11.5 started
INFO:     Starting nodes Xray core
INFO:     Application startup complete.
INFO:     Uvicorn running on https://0.0.0.0:10000 (Press CTRL+C to quit)
INFO:     127.0.0.1:52670 - "POST /api/admin/token HTTP/1.1" 200 OK
INFO:     127.0.0.1:52676 - "GET /api/hosts HTTP/1.1" 200 OK
INFO:     127.0.0.1:52686 - "PUT /api/hosts HTTP/1.1" 200 OK
```