@v-kamerdinerov 2024-11-08
> Can you try to upgrade python version too?

Yep, why not. Which do you prefer? 3.11?