@v-kamerdinerov 2024-11-09
> i try ur docker image and i found something you didn't copy geofies
> 
> ```
> Xray 24.11.5 (Xray, Penetrates Everything.) afc7ec5 (go1.23.2 linux/amd64)
> 
> A unified platform for anti-censorship.
> 2024/11/09 14:47:43 [Info] infra/conf/serial: Reading config: &{Name:stdin: Format:json}
> Failed to start: main: failed to load config files: [stdin:] > infra/conf: invalid field rule > infra/conf: failed to load GeoIP: private > infra/conf: failed to load file: geoip.dat > infra/conf: failed to open file: geoip.dat > open /usr/local/share/xray/geoip.dat: no such file or directory
> 
> ```

Yes, it is possible - since I am using a multi stage build.

I will complete this copying along with adding python 3.12. testing now.