@v-kamerdinerov 2024-11-09
> Run it on a new machine so you can find all bugs


Greetings again.


With the last commit, I successfully updated python to version 3.12. I also updated the libs for compatibility.

```
ubuntu@main:~$ docker exec -ti 1b5a7e5c6510 python --version
Python 3.12.7
```

Also i added the forgotten geosite/geoip db files, thank you for noticing.

```
ubuntu@main:~$ docker exec -ti 1b5a7e5c6510 ls /usr/local/share/xray
geoip.dat geosite.dat
```

Tested everything on a clean install from scratch.

New screenshots showing operability are presented in the original post to the pull request.