@v-kamerdinerov 2024-11-12
> I test this and it working
> However we need @SaintShit opinion

@SaintShit Hello. Can you check this out?